+++ enclone_main/src/lib.rs
@@ -4,5 +4,6 @@ pub mod determine_ref;
 pub mod doublets;
 pub mod inconsistent;
 pub mod main_enclone;
+pub mod merge_onesies;
 pub mod setup;
 pub mod split_orbits;
